Refactoring code: Rearranged classes into separate packages based on concerns.
- Rearranged most of the classes into separate packages.
- Removed printing statements from `AuthenticationService.java` and `JwtAuthFilter.java`.
- Moved `Role.java` to be an inner enum class in `User.java`
