+++ src/main/java/com/anas/meetup/payload/AuthResponse.java
@@ -1,0 +1,21 @@
+package com.anas.meetup.payload;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthResponse {
+
+    private String token;
+    private String email;
+    private String message;
+    private String httpCode;
+
+
+
+}
