--- src/main/java/com/anas/meetup/user/Role.java
@@ -1,6 +1,0 @@
-package com.anas.meetup.user;
-
-public enum Role {
-
-    USER,ADMIN
-}
